requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.30.2
botbuilder-core==4.16.1
botbuilder-integration-aiohttp==4.16.1
botbuilder-schema==4.16.1
botframework-connector==4.16.1
botframework-streaming==4.16.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.66.0
grpcio-status==1.66.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.7
isodate==0.6.1
jiter==0.5.0
jsonpatch==1.33
jsonpickle==1.4.2
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-google-vertexai==1.0.10
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.99
marshmallow==3.21.3
msal==1.30.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
ngrok==1.4.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.41.0
orjson==3.10.7
packaging==24.1
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.19
yarl==1.9.4